+++ tests/test_suite.h
@@ -26,5 +26,6 @@
 int test_cbor(void);
 int test_cose(void);
 int test_interop(void);
+int test_psa_attestation(void);
 
 #endif /* WOLFCOSE_TEST_SUITE_H */